Extract renderSelectOptions out
